+++ src/main/java/IDrawable/Drawable.java
@@ -1,6 +1,6 @@
 package IDrawable;
 
 public interface Drawable {
-    boolean draw(int[][] Canvas);
+    boolean draw(byte[][] Canvas);
 }
